% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bar_graph.R
\name{bar_graph}
\alias{bar_graph}
\title{A function  that plots a a bar graph with different variations}
\usage{
bar_graph(df, xvar, yvar, ascending = FALSE, ordered = FALSE, errorbar = FALSE, xlabel = "", ylabel = "", title = "", flip = FALSE, angle = 0, sizee = 12, sizelabel = 4, dcm = 0, ylimm = c(0, NA))
}
\arguments{
\item{df}{A data frame with x variable, y variable}

\item{xvar}{the variable that will form the x-axis}

\item{yvar}{the variable that will form the y-axis}

\item{ascending}{Boolean deciding the sorting order of the bars.Default:FALSE}

\item{ordered}{Boolean deciding whether bars should be sorted or not sorted.Default:FALSE}

\item{errorbar}{Boolean deciding whether to plot error bars of not. Default:FALSE}

\item{xlabel}{the x label of the plot}

\item{ylabel}{the y label of the plot}

\item{title}{the title of the plot}

\item{flip}{Boolean deciding whether to flip the bar plot or not. Default:FALSE}

\item{angle}{the angle position of the lables}

\item{sizee}{the size of title}

\item{sizelabel}{the size of labels}

\item{dcm}{the number of decimal numbers}

\item{ylimm}{Set the ylim scale}
}
\description{
A function  that plots a a bar graph with different variations
}
\examples{


}
