% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/boxplot.R
\name{boxplot}
\alias{boxplot}
\title{Function to perform simple boxplot}
\usage{
boxplot(
  df,
  variable,
  ylablabel = " ",
  xlablabel = "",
  title = "",
  angle = 0,
  sizee = 12
)
}
\arguments{
\item{df}{A data frame}

\item{variable}{the variable that will form the x-axis}

\item{ylablabel}{the y label of the plot}

\item{xlablabel}{the x label of the plot}

\item{title}{the title of the plot}

\item{angle}{the of lables}

\item{sizee}{the size of title}
}
\description{
Function to perform simple boxplot
}
\examples{
boxplot(df = diamonds, variable = price )

}
