% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/histogram.R
\name{histogram_summaries}
\alias{histogram_summaries}
\title{Function to perform simple barplot}
\usage{
histogram_summaries(data1, column, title = "", xlablabel = "")
}
\arguments{
\item{data1}{A data frame}

\item{column}{the variable that will form the x-axis}

\item{title}{the title of the plot}

\item{xlablabel}{the x label of the plot}

\item{sizee}{the size of title}

\item{angle}{the of lables}
}
\description{
Function to perform simple barplot
}
\examples{
boxplot(df = diamonds, variable = price )

}
