% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/install_load_packages.R
\name{load_install_packages}
\alias{load_install_packages}
\title{A function  that loads installed  packages and loads and installs those not installed.}
\usage{
load_install_packages(x)
}
\arguments{
\item{x}{A vector of all the packages you want to install or load}
}
\description{
A function  that loads installed  packages and loads and installs those not installed.
}
\examples{
load_install_packages(c("tidyverse","knitr"))

}
