% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summarise_by_group.R
\name{summarise_by_group}
\alias{summarise_by_group}
\title{A function to display level of missingness in a data frame provided}
\usage{
summarise_by_group(df, variable, group_variable, dcm = 0)
}
\arguments{
\item{df}{A data frame containing the variables to be summarised}

\item{group_variable}{The grouping variable factor}

\item{dcm}{Number of decimal places to be previewed}
}
\value{
A data frame structure showing the freq distribution
}
\description{
A function to display level of missingness in a data frame provided
}
\examples{
 if(!require(ggplot2)){install.packages(ggplot2)};summary_group(diamonds, variable = clarity,group_variable = cut)

}
