% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tabplot_bar.R
\name{tab_to_df}
\alias{tab_to_df}
\title{A Function that converts a table to data.frame.}
\usage{
tab_to_df(tab, sort_values = "normal", add_totals = TRUE, dcm_place = 0)
}
\arguments{
\item{tab}{A table summary generated eg table of variable x}

\item{sort_values}{Sorts by Count column.Options;"normal","desc" and "asc". Default is "normal"}

\item{add_totals}{Adds additional row of Totals. Default : TRUE}

\item{dcm_place}{Decides the number of decimal places. Default: 0}
}
\value{
A data frame structure of the table provided
}
\description{
A Function that converts a table to data.frame.
}
\examples{
tab_to_df(table(CO2$Type))

}
