% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/barplot_three_way_percent.R
\name{threeway_barplot_percent}
\alias{threeway_barplot_percent}
\title{A function  that plots a three way barplot with labels (Percentages) see example}
\usage{
threeway_barplot_percent(
  df,
  x_variable,
  y_variable,
  groupby_variable,
  title,
  xlablabel = "",
  ylablabel = "Percent",
  scale_values = c(),
  sizee = 12,
  angle = 0,
  sizelabel = 3
)
}
\arguments{
\item{df}{A data frame with x variable, y variable and the factor variable.}

\item{x_variable}{the variable that will form the x-axis}

\item{y_variable}{the variable that will form the y-axis}

\item{groupby_variable}{the factor variable that holds the groupings}

\item{title}{the title of the plot}

\item{xlablabel}{the x label of the plot
@param ylablabel the y label of the plot
@param scale_values Specify the factor groups with the colours you would like them to take eg "MMale"  =  "#66AAD7"
@param sizee the size of title
@param angle the of lables
@param sizelabel the size of labels}
}
\description{
A function  that plots a three way barplot with labels (Percentages) see example
}
\examples{


}
